# backend/python-flask-server/swagger_server/models/contact.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class Contact(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None, mail: str=None, role: str=None):
        """
        Contact - a model defined in Swagger

        :param id: The id of this Contact.
        :type id: int
        :param name: The name of this Contact.
        :type name: str
        :param mail: The mail of this Contact.
        :type mail: str
        :param role: The role of this Contact.
        :type role: str
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'mail': str,
            'role': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'mail': 'mail',
            'role': 'role'
        }

        self._id = id
        self._name = name
        self._mail = mail
        self._role = role

    @classmethod
    def from_dict(cls, dikt) -> 'Contact':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Contact of this Contact.
        :rtype: Contact
        """
        return deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """
        Gets the id of this Contact.
        internal id of contact, gets created by system

        :return: The id of this Contact.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """
        Sets the id of this Contact.
        internal id of contact, gets created by system

        :param id: The id of this Contact.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """
        Gets the name of this Contact.
        name of contact

        :return: The name of this Contact.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """
        Sets the name of this Contact.
        name of contact

        :param name: The name of this Contact.
        :type name: str
        """

        self._name = name

    @property
    def mail(self) -> str:
        """
        Gets the mail of this Contact.
        mail adress of contact

        :return: The mail of this Contact.
        :rtype: str
        """
        return self._mail

    @mail.setter
    def mail(self, mail: str):
        """
        Sets the mail of this Contact.
        mail adress of contact

        :param mail: The mail of this Contact.
        :type mail: str
        """

        self._mail = mail

    @property
    def role(self) -> str:
        """
        Gets the role of this Contact.
        role of this contact

        :return: The role of this Contact.
        :rtype: str
        """
        return self._role

    @role.setter
    def role(self, role: str):
        """
        Sets the role of this Contact.
        role of this contact

        :param role: The role of this Contact.
        :type role: str
        """

        self._role = role