# backend/python-flask-server/swagger_server/models/customer.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class Customer(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None):
        """
        Customer - a model defined in Swagger

        :param id: The id of this Customer.
        :type id: int
        :param name: The name of this Customer.
        :type name: str
        """
        self.swagger_types = {
            'id': int,
            'name': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name'
        }

        self._id = id
        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> 'Customer':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Customer of this Customer.
        :rtype: Customer
        """
        return deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """
        Gets the id of this Customer.
        internal id of customer, gets created by system

        :return: The id of this Customer.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """
        Sets the id of this Customer.
        internal id of customer, gets created by system

        :param id: The id of this Customer.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """
        Gets the name of this Customer.
        name of the customer

        :return: The name of this Customer.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """
        Sets the name of this Customer.
        name of the customer

        :param name: The name of this Customer.
        :type name: str
        """

        self._name = name